+++ apps/api/src/modules/project/schema/project-create.schema.ts
@@ -1,5 +1,6 @@
 import {Static, Type} from '@sinclair/typebox';
 import {Nullable} from '../../../common/typebox/nullable.js';
+import {SnippetTerminalBorderType} from './project.schema.js';
 
 export const SnippetFrameCreateRequestSchema = Type.Object(
   {
@@ -45,6 +46,7 @@ export const SnippetTerminalCreateRequestSchema = Type.Object(
     opacity: Nullable(Type.Number({minimum: 0, maximum: 100, default: 100})),
     showHeader: Type.Boolean(),
     showWatermark: Nullable(Type.Boolean({default: true})),
+    borderType: Nullable(SnippetTerminalBorderType),
   },
   {title: 'SnippetTerminalCreateRequest'},
 );