+++ apps/codeimage/src/state/editor/model.ts
@@ -61,6 +61,8 @@ export interface TerminalState {
   showGlassReflection: boolean;
   opacity: number;
   alternativeTheme: boolean;
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  borderType: ('glass' | (string & {})) | null;
 }
 
 export type PersistedTerminalState = Pick<
@@ -75,6 +77,7 @@ export type PersistedTerminalState = Pick<
   | 'showGlassReflection'
   | 'opacity'
   | 'alternativeTheme'
+  | 'borderType'
 >;
 
 export interface ProjectEditorPersistedState {