+++ packages/dom-export/src/lib/clonePseudoElements.ts
@@ -1,3 +1,4 @@
+import {isIOS, isSafari} from '@solid-primitives/platform';
 import {toArray, uuid} from './util';
 
 type Pseudo = ':before' | ':after';
@@ -24,6 +25,9 @@ function getPseudoElementStyle(
   style: CSSStyleDeclaration,
 ): Text {
   const selector = `.${className}:${pseudo}`;
+  if (isIOS || isSafari) {
+    style.boxShadow = 'unset';
+  }
   const cssText = style.cssText
     ? formatCSSText(style)
     : formatCSSProperties(style);