+++ apps/api/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @codeimage/api
 
+## 0.4.1
+
+### Patch Changes
+
+- [#631](https://github.com/riccardoperra/codeimage/pull/631) [`d8b472a2`](https://github.com/riccardoperra/codeimage/commit/d8b472a2a3ddc1fd113d89688d045755046ee8b0) Thanks [@riccardoperra](https://github.com/riccardoperra)! - feat: add support for go lang
+
+## 0.4.0
+
+### Minor Changes
+
+- [#611](https://github.com/riccardoperra/codeimage/pull/611) [`e94e8a2d`](https://github.com/riccardoperra/codeimage/commit/e94e8a2d5ec01cddd7593c837bf59f55b8a4e231) Thanks [@riccardoperra](https://github.com/riccardoperra)! - feat(app,api) add new border type glass option
+
+- [#624](https://github.com/riccardoperra/codeimage/pull/624) [`9e537bcd`](https://github.com/riccardoperra/codeimage/commit/9e537bcdf92ffaf43eb5e38f5f7fb578eaef9709) Thanks [@riccardoperra](https://github.com/riccardoperra)! - feat(app,api): add line number start option
+
 ## 0.3.2
 
 ### Patch Changes