+++ packages/highlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @codeimage/highlight
 
+## 0.10.9
+
+### Patch Changes
+
+- [#631](https://github.com/riccardoperra/codeimage/pull/631) [`d8b472a2`](https://github.com/riccardoperra/codeimage/commit/d8b472a2a3ddc1fd113d89688d045755046ee8b0) Thanks [@riccardoperra](https://github.com/riccardoperra)! - feat: add support for go lang
+
 ## 0.10.8
 
 ### Patch Changes